Add options to backfills

+++ spec/spec_helper.rb
@@ -22,9 +22,17 @@ RSpec.configure do |config|
   end
 
   config.before :each do
-    User.delete_all
-    Employee.delete_all
-    DataDrip::BackfillRunBatch.delete_all
-    DataDrip::BackfillRun.delete_all
+    if defined?(DataDrip::BackfillRunBatch)
+      DataDrip::BackfillRunBatch.delete_all
+    end
+    DataDrip::BackfillRun.delete_all if defined?(DataDrip::BackfillRun)
+    Employee.delete_all if defined?(Employee)
+    User.delete_all if defined?(User)
+
+    if ActiveRecord::Base.connection.adapter_name == "SQLite"
+      ActiveRecord::Base.connection.execute(
+        "DELETE FROM sqlite_sequence WHERE name IN ('users', 'employees', 'data_drip_backfill_runs', 'data_drip_backfill_run_batches')"
+      )
+    end
   end
 end
